Add script to add an entry in dictionaries (default gesture: NVDA+alt+j) + Code refactoring
